refactor Web3Inbox component, add dynamic import
