+++ src/components/common/Layout/components/Account/components/Web3Inbox/index.ts
@@ -1,0 +1,3 @@
+import Web3Inbox from "./Web3Inbox"
+
+export default Web3Inbox
